universal/rs: reorganize
