+++ universal/rs/messages/src/types/output_token.rs
